backend/saas/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.77.0
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
stability-sdk==0.8.6
starlette==0.46.2
aiohttp==3.9.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
unidecode==1.4.0
urllib3==2.4.0
uvicorn==0.23.2
watchgod==0.8.2
PyJWT==2.8.0
email-validator==2.1.0
